# openapi-spec.yaml
openapi: 3.0.0
paths:
...
  /search/ai:
    post:
      description: Parses a free-form booking request, generates semantic search
        embeddings, and returns ranked service matches.
      operationId: AiSearchController_search
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSearchRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSearchResponseDto"
      summary: Search services using natural language
      tags:
        - AiSearch
...
info:
  title: Glamlink API
  description: The Glamlink API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    AiSearchRequestDto:
      type: object
      properties:
        query:
          type: string
          example: I need a hairstylist for knotless braids tomorrow morning around East
            Legon.
        limit:
          type: number
          minimum: 1
          maximum: 50
          default: 10
          example: 10
        offset:
          type: number
          minimum: 0
          default: 0
          example: 0
        parsedQueryToken:
          type: string
          description: Opaque token returned from a previous AI search to reuse parsed
            query context.
          example: eyJzZXJ2aWNlSW50ZW50Ijoia25vdGxlc3MgYnJhaWRzIn0=
      required:
        - query
    TimeRangeDto:
      type: object
      properties:
        start:
          type: number
          example: 480
        end:
          type: number
          example: 720
      required:
        - start
        - end
    ParsedQueryDto:
      type: object
      properties:
        serviceIntent:
          type: string
          example: knotless braids
        location:
          type: object
          example: East Legon
          nullable: true
        dateTime:
          type: object
          example: tomorrow morning
          nullable: true
        resolvedDate:
          type: object
          example: 2026-03-14
          nullable: true
        timeRange:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/TimeRangeDto"
      required:
        - serviceIntent
    AiSearchResultDto:
      type: object
      properties:
        id:
          type: string
          description: Service ID
        vendorId:
          type: string
        vendorName:
          type: string
        vendorHandle:
          type: string
        vendorAvatar:
          type: object
          nullable: true
        serviceName:
          type: string
        description:
          type: object
          nullable: true
        priceCents:
          type: number
          example: 35000
        durationMinutes:
          type: number
          example: 90
        distanceKm:
          type: object
          nullable: true
          example: 4.2
        rating:
          type: number
          example: 4.8
        reviewCount:
          type: number
          example: 122
        similarityScore:
          type: number
          example: 38
          description: Similarity score as a percentage from 0 to 100.
        availableSlots:
          example:
            - 9:00 AM
            - 9:30 AM
            - 10:00 AM
          type: array
          items:
            type: string
        categoryName:
          type: object
          nullable: true
        vendorLocation:
          type: object
          nullable: true
        vendorTitle:
          type: object
          nullable: true
        yearsExperience:
          type: object
          nullable: true
          example: 6
        bookingCount:
          type: number
          example: 240
        depositPercent:
          type: object
          nullable: true
          example: 20
        includes:
          example:
            - Hot wash
            - Hair trim
          type: array
          items:
            type: string
        serviceImage:
          type: object
          nullable: true
      required:
        - id
        - vendorId
        - vendorName
        - vendorHandle
        - serviceName
        - priceCents
        - durationMinutes
        - rating
        - reviewCount
        - similarityScore
        - availableSlots
        - bookingCount
        - includes
    AiSearchResponseDto:
      type: object
      properties:
        parsedQuery:
          $ref: "#/components/schemas/ParsedQueryDto"
        parsedQueryToken:
          type: string
          description: Opaque token the client can send back on subsequent paginated
            requests.
          example: eyJzZXJ2aWNlSW50ZW50Ijoia25vdGxlc3MgYnJhaWRzIiwibG9jYXRpb24iOiJFYXN0IExlZ29uIn0=
        results:
          type: array
          items:
            $ref: "#/components/schemas/AiSearchResultDto"
        totalEstimate:
          type: number
          example: 24
        hasMore:
          type: boolean
          example: true
      required:
        - parsedQuery
        - parsedQueryToken
        - results
        - totalEstimate
        - hasMore
...